+++ CHANGELOG.rst
@@ -2,7 +2,8 @@ v0.106
 ------
 
   * Move  ``station_location`` to station level 
-  * Change instrument level in network file to accept a list of intruments
+  * Change 'instrument' in network file to 'instruments': list of 'instrument'
+    for OBSs, there will be only one element
   * add Comment in format station XML
   * add start_time and end_time to channel, if not given they take station dates'
   * add iris_unit list to input_unit and output_unit  field